% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{add_sd}
\alias{add_sd}
\title{Add standard deviations to BSD}
\usage{
add_sd(bsd_means, stdev, stdev_range = NULL)
}
\arguments{
\item{bsd_means}{vector of BSD means}

\item{stdev}{stdev proportion to BSD mean value}

\item{stdev_range}{range of std coefficients to draw from (uniform)}
}
\value{
dataframe of mean sizes + scaled sds
}
\description{
Add standard deviations to BSD
}
