% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{AICc}
\alias{AICc}
\title{AICc (from LDATS)
Directly copied from weecology/LDATS because installing LDATS is weird on Travis.}
\usage{
AICc(object)
}
\arguments{
\item{object}{model}
}
\value{
AICc
}
\description{
AICc (from LDATS)
Directly copied from weecology/LDATS because installing LDATS is weird on Travis.
}
