% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{assign_ind_sizes}
\alias{assign_ind_sizes}
\title{Assign sizes to all individuals of a species given abundance, mean, and sd
Assumes a normal distribution}
\usage{
assign_ind_sizes(bsd_index, bsd)
}
\arguments{
\item{bsd_index}{row of bsd - corresponds to species}

\item{bsd}{bsd}
}
\value{
data frame of \code{individual_species_ids} and \code{individual_sizes} for a single species
}
\description{
Assign sizes to all individuals of a species given abundance, mean, and sd
Assumes a normal distribution
}
