% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{combine_abds}
\alias{combine_abds}
\title{Combine BSD and SAD to produce ISD
Assign abundances to body sizes
Current: randomly combine one time
Possible future: incoporate size-abundance relationship
Possible future: use all possible combinations}
\usage{
combine_abds(sad, bsd)
}
\arguments{
\item{sad}{draw from METE SAD (draw_sad)}

\item{bsd}{draw from uniform bsd (draw_bsd)}
}
\value{
bsd with \code{abundance} column
}
\description{
Combine BSD and SAD to produce ISD
Assign abundances to body sizes
Current: randomly combine one time
Possible future: incoporate size-abundance relationship
Possible future: use all possible combinations
}
