% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{community_overlap}
\alias{community_overlap}
\title{Community level overlap}
\usage{
community_overlap(community_df)
}
\arguments{
\item{community_df}{full community df}
}
\value{
df of overlap values for all species pairs + harmonic mean of abundances
}
\description{
Community level overlap
}
