% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_ascii.R
\name{convert_to_ascii}
\alias{convert_to_ascii}
\title{Converts Vector to ASCII}
\usage{
convert_to_ascii(x)
}
\arguments{
\item{x}{A vector of strings}
}
\value{
A vector of strings converted to ASCII encoding.
}
\description{
This is a utility function that converts vectors into ASCII encoding. Please see the example below to see how to apply this function to a data frame. This function's purpose is to ensure that the data frame included in this package is in ASCII encoding, which will make it more accessible across machines and operating systems.
}
\examples{
convert_to_ascii(c("café", "jalapeño"))

test <- data.frame(
names = c("café", "jalapeño"),
more_names = c("café", "jalapeño"),
even_more_names = c("café", "jalapeño")
)

test_df <- lapply(test, function(x) if(is.character(x)) convert_to_ascii(x) else x)
test_df <- as.data.frame(test_df)

}
