% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_methods.R
\name{count_gaps}
\alias{count_gaps}
\title{Count gaps}
\usage{
count_gaps(integrated_density)
}
\arguments{
\item{integrated_density}{result of find_gaps(integrated_density)}
}
\value{
number of gaps
}
\description{
Count gaps
}
