% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_methods.R
\name{count_peaks}
\alias{count_peaks}
\title{Count peaks}
\usage{
count_peaks(integrated_density)
}
\arguments{
\item{integrated_density}{result of find_gaps(integrated_density).}
}
\value{
number of peaks
}
\description{
Count peaks
}
