% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_list.R
\name{create_data_list}
\alias{create_data_list}
\title{Create a list of data objects}
\usage{
create_data_list(exclude_objects)
}
\arguments{
\item{exclude_objects}{A character vector of objects to exclude from the list. Make sure to always exclude the full data frame into which you'll be appending the new data. You should also exclude the \code{exclude_objects} vector itself, as well as any other objects that are not the data frames you intend to pull into the list. Default is \code{c("full_edit_data", "exclude_objects")}.}
}
\value{
A list of data frames that are in the global environment, excluding the objects specified in \code{exclude_objects}. This function serves as a quick way to pull the data frames without having to manually type out each one. You will also see a new object in your global environment called \code{included_objects}, which is a character vector of the objects that were included in the list. This is useful later, when you run the \code{iterated_references} function.
}
\description{
Create a list of data objects
}
\examples{
\dontrun{exclude_objects <- c("full_edit_data", "exclude_objects")}
\dontrun{data_list <- create_data_list(exclude_objects)}
}
