% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{draw_bsd}
\alias{draw_bsd}
\title{Draw BSD from uniform}
\usage{
draw_bsd(community_pars)
}
\arguments{
\item{community_pars}{result of get_sim_pars or get_community_pars}
}
\value{
vector of means for BSD drawn from runif
}
\description{
Draw BSD from uniform
}
