% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{draw_sad}
\alias{draw_sad}
\title{Draw SAD from logseries given S and N}
\usage{
draw_sad(community_pars)
}
\arguments{
\item{community_pars}{result of get_sim_pars}
}
\value{
draw from the logseries SAD predicted by METE
}
\description{
Draw SAD from logseries given S and N
}
