% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{draw_sim}
\alias{draw_sim}
\title{Draw a sim}
\usage{
draw_sim(community_pars, sim_index = 1)
}
\arguments{
\item{community_pars}{result of get_sim_pars}

\item{sim_index}{for drake}
}
\value{
dataframe of \code{species} and \code{wgt} drawn using \code{community_pars}
}
\description{
Draw a sim
}
