% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{draw_uniform_bsd}
\alias{draw_uniform_bsd}
\title{Draw a BSD from uniform}
\usage{
draw_uniform_bsd(s, min, max)
}
\arguments{
\item{s}{Number of species}

\item{min}{Min size}

\item{max}{Max size}
}
\value{
vector of length s, species mean body sizes
}
\description{
(Really a wrapper for \code{sample})
}
