% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{draw_unimodal_bsd}
\alias{draw_unimodal_bsd}
\title{Draw a BSD from unimodal distribution}
\usage{
draw_unimodal_bsd(emp_vector)
}
\arguments{
\item{emp_vector}{Vector of mean body sizes to base kde on}
}
\value{
Vector of body sizes drawn from a smooth probability dist based on kde based on emp_vector
}
\description{
Unimodal distribution based on kde of empirical vector
After Allen 1999
}
