% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{edge_proportion}
\alias{edge_proportion}
\title{Get edge proportion}
\usage{
edge_proportion(overlap_df, weight_col = NULL)
}
\arguments{
\item{overlap_df}{df of overlap vals}

\item{weight_col}{"hm", "total_n", "prod_n", NULL}
}
\value{
prop greater than .7 less than .2
}
\description{
Get edge proportion
}
