% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_methods.R
\name{find_gaps}
\alias{find_gaps}
\title{Find gaps to a given threshold}
\usage{
find_gaps(threshold, integrated_density)
}
\arguments{
\item{threshold}{cutoff probability for gaps}

\item{integrated_density}{result of get_integrated_density}
}
\value{
integrated_density with columns for \code{is_gap_start}, \code{is_gap_end}
}
\description{
Cuts integrated density off at a given probability and finds gaps in the remaining distribution. (Gaps defined as non-sequential starting values, as would occur if the intervening values got filtered out when the distribution was cut off).
}
