% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{find_gaps_scaled}
\alias{find_gaps_scaled}
\title{Find gaps in bsd}
\usage{
find_gaps_scaled(bsd)
}
\arguments{
\item{bsd}{vector of species means}
}
\value{
vector of length length(bsd) - 1, ratios of means
}
\description{
Find gaps in bsd
}
