% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_edit_data.R
\docType{data}
\name{full_edit_data}
\alias{full_edit_data}
\title{Full EDIT Data Set}
\format{
A data frame with 3951 rows and 9 variables.
\describe{
\item{Name}{Name of the IIA document (character)}
\item{Type}{Type of IIA document (factor)}
\item{Date.of.signature}{Date when the IIA was signed (Date)}
\item{Date.of.entry.in.force}{Date when the IIA entered into force (Date)}
\item{Date.of.termination}{Date when the IIA was terminated (Date)}
\item{Date.of.publication}{Date when the IIA was published (Date)}
\item{Origin.language}{Languages of the IIA document (character)}
\item{Party.ISO.codes}{Aggregated ISO codes of the parties involved (character)}
\item{Doc.Reference}{Document identifier (character)}
}
}
\source{
This data set is sourced from the \href{https://edit.wti.org/document/investment-treaty/search}{EDIT website}, where comprehensive IIA data sets can be accessed.
}
\usage{
full_edit_data
}
\description{
This data set contains information on International Investment Agreements (IIAs), including basic identifying details and key dates.
}
\examples{
data(full_edit_data)
head(full_edit_data)
}
\keyword{datasets}
