% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{generate_sim_draws}
\alias{generate_sim_draws}
\title{draw sims all at once}
\usage{
generate_sim_draws(base_dat, dat_name = NULL, fixed_sds, sd_range, nsim)
}
\arguments{
\item{base_dat}{data}

\item{dat_name}{to add to pars}

\item{fixed_sds}{sd pars}

\item{sd_range}{range}

\item{nsim}{nb sim}
}
\value{
list of sims
}
\description{
draw sims all at once
}
