% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{get_community_pars}
\alias{get_community_pars}
\title{Get community parameters}
\usage{
get_community_pars(raw_dat)
}
\arguments{
\item{raw_dat}{community data frame with columns \code{species}', \code{wgt}}
}
\value{
list of community parameters for sim draws: \code{S} = number of species, \code{N} = number of individuals, \code{min_size} = minimum body size, \code{max_size} = maximum body size
}
\description{
Get community parameters
}
