% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_methods.R
\name{get_datname}
\alias{get_datname}
\title{Get dataset name
For collecting results.}
\usage{
get_datname(integrated_density)
}
\arguments{
\item{integrated_density}{result of find_gaps(integrated_density)}
}
\value{
Name of dataset used in integrated_density/sims
}
\description{
Get dataset name
For collecting results.
}
