% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd_methods.R
\name{get_id}
\alias{get_id}
\title{Get id from community data}
\usage{
get_id(dataset)
}
\arguments{
\item{dataset}{list of community data, pars}
}
\value{
list of id, pars
}
\description{
Get id from community data
}
