% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd_methods.R
\name{get_id_wrapped}
\alias{get_id_wrapped}
\title{Wrapper for get_id}
\usage{
get_id_wrapped(datasets)
}
\arguments{
\item{datasets}{list of datasets}
}
\value{
list of ids
}
\description{
Wrapper for get_id
}
