% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{get_mean_difference}
\alias{get_mean_difference}
\title{Get mean difference between modes}
\usage{
get_mean_difference(integrated_density)
}
\arguments{
\item{integrated_density}{integrated_density}
}
\value{
mean difference
}
\description{
Get mean difference between modes
}
