% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{get_mean_p}
\alias{get_mean_p}
\title{Get mean probability density}
\usage{
get_mean_p(integrated_density)
}
\arguments{
\item{integrated_density}{integrated_density}
}
\value{
mean
}
\description{
Get mean probability density
}
