% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{get_mode_ratio}
\alias{get_mode_ratio}
\title{Get mean ratio or stdev of ratio modes}
\usage{
get_mode_ratio(integrated_density, what = "mean")
}
\arguments{
\item{integrated_density}{integrated_density}

\item{what}{"mean" or "stdev"}
}
\value{
mean or stdev
}
\description{
Get mean or stdev of ratio of modes
}
