% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{get_n_clumps}
\alias{get_n_clumps}
\title{Get nb clumps}
\usage{
get_n_clumps(mean_size_vector)
}
\arguments{
\item{mean_size_vector}{bsd}
}
\value{
nb clumps identified as nb Gs in GMM
}
\description{
Get nb clumps
}
