% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{get_n_modes}
\alias{get_n_modes}
\title{Get nb modes}
\usage{
get_n_modes(size_vector, max_nb_clumps = 4, measurer = "aicc")
}
\arguments{
\item{size_vector}{bsd}

\item{max_nb_clumps}{max nb modes to try}

\item{measurer}{"aicc" or "bic"}
}
\value{
nb modes identified as nb turnpoints in best GMM via selector function
}
\description{
Get nb modes
}
