% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{get_nind_permode}
\alias{get_nind_permode}
\title{Get nb individuals  in each mode}
\usage{
get_nind_permode(size_vector, max_nb_gaussians = 8, measurer = "bic")
}
\arguments{
\item{size_vector}{isd}

\item{max_nb_gaussians}{nb modes to use}

\item{measurer}{"aicc" or "bic"}
}
\value{
nb modes identified as nb turnpoints in best GMM via selector function
}
\description{
Get nb individuals  in each mode
}
