% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_methods.R
\name{get_result}
\alias{get_result}
\title{Collect results}
\usage{
get_result(integrated_density)
}
\arguments{
\item{integrated_density}{Result of find_gaps}
}
\value{
data frame 1 row long with summary of thresholding exercise.
}
\description{
Collect results
}
