% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{get_scaled_e_change}
\alias{get_scaled_e_change}
\title{Get elevation change}
\usage{
get_scaled_e_change(integrated_density)
}
\arguments{
\item{integrated_density}{integrated_density}
}
\value{
scaled e change
}
\description{
Get elevation change
}
