% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{get_species_pairs}
\alias{get_species_pairs}
\title{Get species pairs}
\usage{
get_species_pairs(pair, community_df)
}
\arguments{
\item{pair}{vector of 2 species ids}

\item{community_df}{community df}
}
\value{
list of the two species wgts
}
\description{
Get species pairs
}
