% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{get_ssq_prop}
\alias{get_ssq_prop}
\title{Get between/total ssq}
\usage{
get_ssq_prop(mean_size_vector, nbclumps = NULL)
}
\arguments{
\item{mean_size_vector}{mean size vector}
}
\value{
btwn/total ssq for kmeans
}
\description{
Get between/total ssq
}
