% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{get_stdev}
\alias{get_stdev}
\title{Retrieve standard deviation}
\usage{
get_stdev(integrated_density)
}
\arguments{
\item{integrated_density}{result of find_gaps(integrated_density)}
}
\value{
sd
}
\description{
Retrieve sd used to create ISD for sims. For collecting results.
}
