% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_methods.R
\name{get_threshold}
\alias{get_threshold}
\title{Get threshold}
\usage{
get_threshold(integrated_density)
}
\arguments{
\item{integrated_density}{result of find_gaps(integrated_density)}
}
\value{
threshold value
}
\description{
Retrieve the threshold probability used in find_gaps. For collecting results.
}
