% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{get_type}
\alias{get_type}
\title{Get type
For collecting results.}
\usage{
get_type(integrated_density)
}
\arguments{
\item{integrated_density}{result of find_gaps(integrated_density)}
}
\value{
"emp" or "sim"
}
\description{
Get type
For collecting results.
}
