% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{get_width}
\alias{get_width}
\title{Get width of size spectrum}
\usage{
get_width(integrated_density)
}
\arguments{
\item{integrated_density}{integrated_density}
}
\value{
width
}
\description{
Get width of size spectrum
}
