% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_references.R
\name{iterate_references}
\alias{iterate_references}
\title{Iterates the \code{mark_references} function through EDIT dataframes list}
\usage{
iterate_references(data, dfs, varnames)
}
\arguments{
\item{data}{Object of class data.frame to which you'll append the new columns.}

\item{dfs}{A list of EDIT dataframes. This list can be more easily generated using the \code{create_data_list} function, which will pull all EDIT dataframes from the global environment.}

\item{varnames}{A character vector of the variable names for your new variables. Note that the \code{create_data_list} function will automatically generate an object called \code{included_objects}, which will have the relevant varaible names for each EDIT dataframe.}
}
\value{
A single EDIT dataframe with all the relevant IIA provisions marked as binary variables.
}
\description{
Iterates the \code{mark_references} function through EDIT dataframes list
}
\examples{
\dontrun{iia_data <- iterate_references(dfs = data_list, varnames = included_objects)}
}
