% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{make_bsd}
\alias{make_bsd}
\title{Make BSD
From community dat with species and wgt columns}
\usage{
make_bsd(community)
}
\arguments{
\item{community}{df with species and wgt columns per individual}
}
\value{
df with species, meanwgt, logwgt columns
}
\description{
Make BSD
From community dat with species and wgt columns
}
