% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_methods.R
\name{make_cp_pipeline}
\alias{make_cp_pipeline}
\title{Make community pars pipeline}
\usage{
make_cp_pipeline(dats_pipeline)
}
\arguments{
\item{dats_pipeline}{pipeline of loading raw data}
}
\value{
pipeline to extract communty pars from raw datasets
}
\description{
Make community pars pipeline
}
