% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_methods.R
\name{make_draw_pipeline}
\alias{make_draw_pipeline}
\title{Make draw sims pipeline}
\usage{
make_draw_pipeline(sp_pipeline, sim_indices)
}
\arguments{
\item{sp_pipeline}{pipeline of community pars including standard deviations}

\item{sim_indices}{for drake}
}
\value{
pipeline to extract draw sims based on community pars
}
\description{
Make draw sims pipeline
}
