% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_methods.R
\name{make_id_pipeline}
\alias{make_id_pipeline}
\title{Make community pars pipeline}
\usage{
make_id_pipeline(community_dat_pipeline, community_type = "sim")
}
\arguments{
\item{community_dat_pipeline}{pipeline to either get raw data or draw sims}

\item{community_type}{"sim" or "empirical"}
}
\value{
pipeline to get integrated density of GMM fit to community data
}
\description{
Make community pars pipeline
}
