% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_methods.R
\name{make_id_plots_pipeline}
\alias{make_id_plots_pipeline}
\title{Make id plots plan}
\usage{
make_id_plots_pipeline(id_pipeline_plan, sim_index = c(1:2))
}
\arguments{
\item{id_pipeline_plan}{rbind of empirical, sims pipelines}

\item{sim_index}{which sim(s) to plot}
}
\value{
plan to plot empirical + 1 of each sim setting for every dataset
}
\description{
Make id plots plan
}
