% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd_methods.R
\name{make_isd}
\alias{make_isd}
\title{Make ISD}
\usage{
make_isd(community_dat)
}
\arguments{
\item{community_dat}{community df with energy & sizeclass columns added}
}
\value{
ISD table
}
\description{
Make ISD
}
