% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{make_kde}
\alias{make_kde}
\title{Make scaled kde}
\usage{
make_kde(sbsd, min_size, max_size, npoints)
}
\arguments{
\item{sbsd}{Wgt vector}

\item{min_size}{minimum range for kde}

\item{max_size}{max range for kde}

\item{npoints}{nb eval points, specify to powers of 2}
}
\value{
scaled probability density
}
\description{
Make scaled kde
}
