% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{make_metadata}
\alias{make_metadata}
\title{Get metadata from pars list}
\usage{
make_metadata(id_list)
}
\arguments{
\item{id_list}{list ofi ntegrated density, pars}
}
\value{
metadata df
}
\description{
Get metadata from pars list
}
