% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{make_results}
\alias{make_results}
\title{Make results table}
\usage{
make_results(id_list)
}
\arguments{
\item{id_list}{list ofi ntegrated densities}
}
\value{
df of rsults
}
\description{
Make results table
}
