% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_methods.R
\name{make_sp_pipeline}
\alias{make_sp_pipeline}
\title{Make sim pars pipeline}
\usage{
make_sp_pipeline(cp_pipeline, stdevs, stdev_range = NULL)
}
\arguments{
\item{cp_pipeline}{pipeline of extracting community pars from raw data}
}
\value{
pipeline to add standard deviations to community pars
}
\description{
Make sim pars pipeline
}
