% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_methods.R
\name{make_summary_plots_pipeline}
\alias{make_summary_plots_pipeline}
\title{Make summary plots pipeline}
\usage{
make_summary_plots_pipeline(dats_pipeline)
}
\arguments{
\item{dats_pipeline}{To get dataset names}
}
\value{
plan for summary plots
}
\description{
Make summary plots pipeline
}
