% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_methods.R
\name{make_thresholds_pipeline}
\alias{make_thresholds_pipeline}
\title{Make thresholds pipeline
Make pipeline to evaluate number of gaps/modes in an integrated density given a particular threshold density}
\usage{
make_thresholds_pipeline(id_pipeline, thresholds_to_try)
}
\arguments{
\item{id_pipeline}{pipeline to make integrated densities}

\item{thresholds_to_try}{thresholds for drake}
}
\value{
pipeline to evaluate at thresholds
}
\description{
Make thresholds pipeline
Make pipeline to evaluate number of gaps/modes in an integrated density given a particular threshold density
}
