% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{pair_harmonic_mean}
\alias{pair_harmonic_mean}
\title{Harmonic mean of abundances}
\usage{
pair_harmonic_mean(species_vects)
}
\arguments{
\item{species_vects}{wgt vectors}
}
\value{
Harmonic mean of abundances
}
\description{
Harmonic mean of abundances
}
