% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{pair_prod_abund}
\alias{pair_prod_abund}
\title{Product abundance of two species}
\usage{
pair_prod_abund(species_vects)
}
\arguments{
\item{species_vects}{wgt vectors}
}
\value{
product n
}
\description{
Product abundance of two species
}
