% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{pair_total_abund}
\alias{pair_total_abund}
\title{Total abundance of two species}
\usage{
pair_total_abund(species_vects)
}
\arguments{
\item{species_vects}{wgt vectors}
}
\value{
total n
}
\description{
Total abundance of two species
}
