% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_dataset_ids}
\alias{plot_dataset_ids}
\title{Plot a bunch of things}
\usage{
plot_dataset_ids(dataset_ids)
}
\arguments{
\item{dataset_ids}{list ofids}
}
\value{
list of plots
}
\description{
Plot a bunch of things
}
