% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_integrated_density}
\alias{plot_integrated_density}
\title{Plot integrated density}
\usage{
plot_integrated_density(integrated_density, threshold_lines = NULL,
  plot_title = NULL)
}
\arguments{
\item{integrated_density}{int d dataframe}

\item{threshold_lines}{vector of values to put lines at to mark thresholds}

\item{plot_title}{title for plot}
}
\value{
plot of integrated density with maxima and minima marked
}
\description{
Plot integrated density
}
