% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_isd_methods_summary}
\alias{plot_isd_methods_summary}
\title{Make summary plots}
\usage{
plot_isd_methods_summary(summary_dataframe, dat_name)
}
\arguments{
\item{summary_dataframe}{result}

\item{dat_name}{dat name for this set of plots}
}
\value{
histogram of n peaks, dotplot of npeaks:ngaps by threshold
}
\description{
Make summary plots
}
