% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_overlap_scenarios.R
\name{plot_overlaps}
\alias{plot_overlaps}
\title{Plot histogram of overlap values}
\usage{
plot_overlaps(overlap_df, weighting_col = NULL, type = "histogram")
}
\arguments{
\item{overlap_df}{df with overlap, optionally a column to weight by}

\item{weighting_col}{optional name of column to weight by}

\item{type}{"histogram" or "ecdf"}
}
\value{
histogram
}
\description{
Plot histogram of overlap values
}
