% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_overlap_scenarios.R
\name{plot_pipeline}
\alias{plot_pipeline}
\title{Plot ISD, SBSD, and overlap}
\usage{
plot_pipeline(community_df, weighting_col = NULL,
  overlap_type = "histogram")
}
\arguments{
\item{community_df}{df}

\item{weighting_col}{weighting col to pass to overlap plot}

\item{overlap_type}{"histogram" or "ecdf"}
}
\value{
list of isd, sbsd, overlap plots
}
\description{
Plot ISD, SBSD, and overlap
}
