% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_overlap_scenarios.R
\name{plot_sbsds}
\alias{plot_sbsds}
\title{Plot species body size distributions}
\usage{
plot_sbsds(community_df)
}
\arguments{
\item{community_df}{df with species, wgt}
}
\value{
sbsd plot
}
\description{
Plot species body size distributions
}
