% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{replicate_sim_calls}
\alias{replicate_sim_calls}
\title{Replicte sim call}
\usage{
replicate_sim_calls(sim_call, nsim)
}
\arguments{
\item{sim_call}{sim call}

\item{nsim}{nb replicates}
}
\value{
list of sims
}
\description{
To draw sims.
}
