% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns2.R
\name{sample_constrained_isd}
\alias{sample_constrained_isd}
\title{Sample constrained isd}
\usage{
sample_constrained_isd(isd, size_relationship = "none")
}
\arguments{
\item{isd}{isd to base on}

\item{size_relationship}{"none" or "negative}
}
\value{
isd
}
\description{
Sample constrained isd
}
