% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns2.R
\name{sample_uniform_isd}
\alias{sample_uniform_isd}
\title{Sample uniform isd}
\usage{
sample_uniform_isd(isd)
}
\arguments{
\item{isd}{isd to base it on}
}
\value{
isd drawn from uniform
}
\description{
Sample uniform isd
}
