% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims_methods.R
\name{add_sd}
\alias{add_sd}
\title{Add standard deviations to BSD}
\usage{
add_sd(bsd_means, stdev)
}
\arguments{
\item{bsd_means}{vector of BSD means}

\item{stdev}{stdev proportion to BSD mean value: numeric, vector, or "norm"}
}
\value{
dataframe of mean sizes + scaled sds
}
\description{
Add standard deviations to BSD
}
