% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_methods.R
\name{count_peaks}
\alias{count_peaks}
\title{Count peaks}
\usage{
count_peaks(integrated_density_list)
}
\arguments{
\item{integrated_density_list}{list}
}
\value{
number of peaks
}
\description{
Count peaks
}
