% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{draw_multimodal_bsd}
\alias{draw_multimodal_bsd}
\title{Draw multimodal BSD}
\usage{
draw_multimodal_bsd(emp_vector, min_buffer_coeff = 1,
  max_buffer_coeff = 1, min_mode_gap_coeff = 1.5, min_sd_coeff = 0.1,
  max_sd_coeff = 1, return_full = TRUE)
}
\arguments{
\item{emp_vector}{base vector}

\item{min_buffer_coeff}{defaults to 1}

\item{max_buffer_coeff}{defaults to 1}

\item{min_mode_gap_coeff}{defaults to 1.5 (larger mode must be at leas 150 percent of smaller mode)}

\item{min_sd_coeff}{defaults to .2}

\item{max_sd_coeff}{defaults to .2}

\item{return_full}{return nmodes and mean_p (TRUE) or just the vector?}
}
\value{
list of vector of mean_sizes length S drawn from multimodal distribution; nmodes; mode_p distribution
}
\description{
Draw multimodal BSD
}
