% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{find_gaps}
\alias{find_gaps}
\title{Find gaps in bsd}
\usage{
find_gaps(bsd)
}
\arguments{
\item{bsd}{vector of species means}
}
\value{
vector of length length(bsd) - 1, gaps between means
}
\description{
Find gaps in bsd
}
