% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsd_decay_methods.R
\name{get_n_clumps}
\alias{get_n_clumps}
\title{Get nb clumps}
\usage{
get_n_clumps(mean_size_vector, max_nb_clumps = 4, measurer = "aicc")
}
\arguments{
\item{mean_size_vector}{bsd}

\item{max_nb_clumps}{max nb clumps to try}

\item{measurer}{"aicc" or "bic"}
}
\value{
nb clumps identified as nb Gs in GMM
}
\description{
Get nb clumps
}
