% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{get_toy_portal_data}
\alias{get_toy_portal_data}
\title{Get toy Portal data
Get an arbitrary year of Portal control plot data - for use in developing & testing methods, not for actual analysis}
\usage{
get_toy_portal_data(years = c(1994, 1995))
}
\arguments{
\item{years}{vector of years to pull}
}
\value{
Portal control rodents
}
\description{
Get toy Portal data
Get an arbitrary year of Portal control plot data - for use in developing & testing methods, not for actual analysis
}
