% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_overlap_scenarios.R
\name{sample_community}
\alias{sample_community}
\title{Sample a community}
\usage{
sample_community(means, sds = NULL, abunds, sd_coefficient = 0.15)
}
\arguments{
\item{means}{Vector of species means}

\item{sds}{Vector of standard deviations}

\item{abunds}{Vector of abundanaces}

\item{sd_coefficient}{if sds is not specified, sds = sd_coefficient * means}
}
\value{
data frame of simulated community
}
\description{
Sample a community
}
